Era branch: era5 contour plot

+++ postwrf.sh
@@ -1,4 +1,4 @@
-#PostWRF Version 1.2 (May 2021)
+#PostWRF Version 1.3 (January 2022)
 #Author: Amirhossein Nikfal <ah.nikfal@gmail.com>, <https://github.com/anikfal>
 
 export postwrf_dir=$(pwd)
@@ -12,6 +12,11 @@ my_exit() {
   cd $postwrf_dir
   rm variablesCN1.txt variablesCN2.txt variablesCN3.txt eqname equnit totalequation.txt variables.txt 2>/dev/null
 }
+
+awk_read_onoff () {
+        awk -v pat=$1 '$0~pat {print $3}' namelist.wrf
+}
+
 while getopts hdf:i option; do
   case $option in
   h)
@@ -22,46 +27,42 @@ while getopts hdf:i option; do
     echo "    -i,  PostWRF version and basic informations"
     ;;
   d) cat modules/readme ;;
-  f) if [[ $(echo $2 | rev | cut -c -3 | rev) == ".nc" ]]; then
-    ncl_filedump $2 | grep "( Time, bottom_top, south_north, west_east" | awk '{print $2}' >.wrfvars
-    ncl_filedump $2 | grep "( Time, bottom_top, south_north_stag, west_east )" | awk '{print $2}' >>.wrfvars
-    ncl_filedump $2 | grep "( Time, south_north, west_east" | awk '{print $2}' >>.wrfvars
-    ncl_filedump $2 | grep "Variable" | grep -v "f" | awk '{print $2}' >>.wrfvars
-    echo -e "\n  List of variables inside $2:"
-    echo ""
-    var_list=$(cat .wrfvars)
-    echo $var_list | sed 's/ /,  /g'
-    rm .wrfvars
-  else
-    ncl_filedump "$2.nc" | grep "( Time, bottom_top, south_north, west_east" | awk '{print $2}' >.wrfvars
-    ncl_filedump "$2.nc" | grep "( Time, bottom_top, south_north_stag, west_east )" | awk '{print $2}' >>.wrfvars
-    ncl_filedump "$2.nc" | grep "( Time, south_north, west_east" | awk '{print $2}' >>.wrfvars
-    ncl_filedump "$2.nc" | grep "Variable" | grep -v "f" | awk '{print $2}' >>.wrfvars
-    echo -e "\n  List of variables inside $2:"
-    echo ""
-    var_list=$(cat .wrfvars)
-    echo $var_list | sed 's/ /,  /g'
-    rm .wrfvars
-  fi ;;
+  f)  if [[ $(echo $2 | awk -F. '{ print $NF }') == "nc" ]]; then
+        echo "  Variable names inside the file $2:"
+        ncl_filedump $2 | awk '/short/ {for(i=2;i<=NF;i++) printf $i" "; print ""}'
+        echo "  Variables' description:"
+        ncl_filedump $2 | awk '/long_name/ {for(i=1;i<=NF;i++) printf $i" "; print ""}' | sed -n '5,$ p'
+        else
+        ncl_filedump "$2.nc" | grep "( Time, bottom_top, south_north, west_east" | awk '{print $2}' >.wrfvars
+        ncl_filedump "$2.nc" | grep "( Time, bottom_top, south_north_stag, west_east )" | awk '{print $2}' >>.wrfvars
+        ncl_filedump "$2.nc" | grep "( Time, south_north, west_east" | awk '{print $2}' >>.wrfvars
+        ncl_filedump "$2.nc" | grep "Variable" | grep -v "f" | awk '{print $2}' >>.wrfvars
+        echo -e "\n  List of variables inside $2:"
+        echo ""
+        var_list=$(cat .wrfvars)
+        echo $var_list | sed 's/ /,  /g'
+        rm .wrfvars
+      fi ;;
   i)
-    echo "  PostWRF Version 1.1 (Oct 2020)"
-    echo "  Programmed by Amirhossein Nikfal <ah.nikfal@gmail.com>, <anik@ut.ac.ir>"
+    echo "  PostWRF Version 1.2 (January 2022)"
+    echo "  Author: Amirhossein Nikfal <ah.nikfal@gmail.com>, <anik@ut.ac.ir>"
     ;;
   esac
 done
 
 if [[ $1 != "-h" && $1 != "-i" && $1 != "-f" && $1 != "-d" && $1 != "-p" ]]; then
   export wrfout=$1
-  export extractonoff=$(cat namelist.wrf | grep Extract_On-Off | awk '{print $3}')
-  export contour_onoff=$(cat namelist.wrf | grep CONTOUR_ON_OFF | awk '{print $3}')
-  export crossonoff=$(cat namelist.wrf | grep CROSSSECTION_ON-OFF | awk '{print $3}')
-  export roseonoff=$(cat namelist.wrf | grep windrose_On-Off | awk '{print $3}')
-  export skewtonoff=$(cat namelist.wrf | grep skewT_On-Off | awk '{print $3}')
-  export domainonoff=$(cat namelist.wrf | grep Domain_On-Off | awk '{print $3}')
-  export statisticalonoff=$(cat namelist.wrf | grep STATISTICAL_DIAGRAMS_ON_OFF | awk '{print $3}')
-  export GEOTIFF_ONOFF=$(cat namelist.wrf | grep Geotiff_ON_OFF | awk '{print $3}')
-  export rttov_onoff=$(cat namelist.wrf | grep RTTOV_On-Off | awk '{print $3}')
-  sumopts=$((extractonoff + contour_onoff + crossonoff + roseonoff + skewtonoff + domainonoff + GEOTIFF_ONOFF + statisticalonoff + rttov_onoff))
+  export extractonoff=$(awk_read_onoff Extract_On-Off)
+  export contour_onoff=$(awk_read_onoff CONTOUR_ON_OFF)
+  export crossonoff=$(awk_read_onoff CROSSSECTION_ON-OFF)
+  export roseonoff=$(awk_read_onoff windrose_On-Off)
+  export skewtonoff=$(awk_read_onoff skewT_On-Off)
+  export domainonoff=$(awk_read_onoff Domain_On-Off)
+  export statisticalonoff=$(awk_read_onoff STATISTICAL_DIAGRAMS_ON_OFF)
+  export GEOTIFF_ONOFF=$(awk_read_onoff Geotiff_ON_OFF)
+  export rttov_onoff=$(awk_read_onoff RTTOV_On-Off)
+  export era_onoff=$(awk_read_onoff ERA5_ON_OFF)
+  sumopts=$((extractonoff + contour_onoff + crossonoff + roseonoff + skewtonoff + domainonoff + GEOTIFF_ONOFF + statisticalonoff + rttov_onoff + era_onoff))
   if [[ $sumopts -gt 1 ]]; then
     echo ""
     echo "  More than one section is activated"
@@ -84,9 +85,13 @@ if [[ $1 != "-h" && $1 != "-i" && $1 != "-f" && $1 != "-d" && $1 != "-p" ]]; the
         exit
       fi
     fi
-    ./modules/main.sh
-    rm eqname equnit 2>/dev/null
-    cd $postwrf_dir/modules
-    rm wrfout_d0* eq* *.txt .AllWRFVariables 2>/dev/null
+    if [[ $era_onoff -ne 1 ]]; then
+      ./modules/main.sh
+      rm eqname equnit 2>/dev/null
+      cd $postwrf_dir/modules
+      rm wrfout_d0* eq* *.txt .AllWRFVariables 2>/dev/null
+    else
+      ./modules_era/main_era.sh
+    fi
   fi
 fi
